improve integration test UX
when running tests outside CI, sbt will try to open a socket connection to the
configured MySQL database. If it can't open the connection, a warning will be
omitted and integration tests will be skipped. In CI, integration tests will
always run. (CI is detected by the presence of a `CI` environment variable set
to `true`.)